requirements-dev.txt
# 開発用
-r requirements.txt

python-dotenv==1.1.1  # 環境変数読み込み

# Lint・フォーマッタ
flake8==7.3.0
black==24.4.2
isort==5.13.2

# ホットリロード用
watchfiles==1.1.0

# CLI補助
ipython==8.24.0

# YAML設定読み込み（設定ファイルで使う場合）
PyYAML==6.0.2

# レートリミット
slowapi==0.1.9
Deprecated==1.2.18
limits==5.4.0
wrapt==1.17.2
